Fix tests for configure command
